+++ core/src/state/mod.rs
@@ -317,7 +317,7 @@ impl<'a> State<'a> {
             std::thread::Builder::new()
                 .name("txpool_update_state".into())
                 .spawn(move || {
-                    txpool_clone.notify_state_start(accounts_for_txpool);
+                    txpool_clone.notify_modified_accounts(accounts_for_txpool);
                 })
                 .expect("can not notify tx pool to start state");
         }
